Changes made to accommodate use by IRODS (eventId must be long since IRODS event id is a timestamp) also need to call the eventParser with an event list rather than a fimename
git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/WDK/trunk@19574 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Model/src/main/java/org/gusdb/wdk/model/user/dataset/event/UserDatasetShareEvent.java
@@ -13,7 +13,7 @@ public class UserDatasetShareEvent extends UserDatasetEvent {
   private Integer userId;
   private ShareAction action;
   
-  UserDatasetShareEvent(Integer eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType, Integer userId, ShareAction action) {
+  UserDatasetShareEvent(Long eventId, Set<String> projectsFilter, Integer userDatasetId, UserDatasetType userDatasetType, Integer userId, ShareAction action) {
     super(eventId, projectsFilter, userDatasetId, userDatasetType);
     this.userId = userId;
     this.action = action;
